refactor(v3): migrate label_utils.py from utils to services layer
Migrate label_utils.py to services layer to better align with its role
as a domain service used by multiple modules (orchestra, roles, services).

Changes:
- Move src/vibe3/utils/label_utils.py to src/vibe3/services/label_utils.py
- Update import paths in 6 consumer files:
  - src/vibe3/roles/governance.py
  - src/vibe3/orchestra/global_dispatch_coordinator.py
  - src/vibe3/orchestra/queue_persistence_mixin.py
  - src/vibe3/orchestra/queue_operations.py
  - src/vibe3/services/scan_service.py
  - src/vibe3/services/task_status_classifier.py
- Update test import path in tests/vibe3/utils/test_label_utils.py
- Remove label_utils references from src/vibe3/utils/README.md

Verification:
- All 16 label_utils tests pass
- No type errors (mypy)
- No lint errors (ruff)
- No circular dependencies (label_utils only depends on clients/models/roles.definitions)

Closes #1268
